Refactor blueprint and schema modules: consolidate structures and remove unused code
